+++ noodles-gff/src/feature.rs
@@ -1,3 +1,5 @@
 //! Feature record.
 
 pub mod record;
+
+pub use self::record::Record;